[Xamarin.Android.Build.Tasks] Fix the AsyncTask so it does not duplicate log entries.
There is a bug in the AsyncTask where if you are already on the UI thread
it will Log the message/error to the logger and still queue the message
for later processing. This results in the log being shown twice!

So we need to exit the various log methods if we are on the UI thread.