[Xamarin.Android.Build.Tasks] Fix the Resource only build.
The UpdateAndroidResources build Task is broken. The refactor
which moved `_ValidatePackageProperties` did not take into
account the face that UpdateAndroidResources does NOT call
`Build`.

This commit fixes this issue by ensuring that `_ValidatePackageProperties`
is called once we have Resolved the SDK which is what we do for every
build regardless.