[msbuild] Add members used by monodroid_create_appdomain to linker whitelist
Related to work done in PR: https://github.com/xamarin/xamarin-android/pull/359

If for some reason the designer starts loading a mscorlib that has been linked
away, we tend to lose the `AppDomain.CreateDomain` function and
`AppDomainSetup.ShadowCopyFiles` property that we use to setup and initialize the
new domain which create a native crash.

Thus this commits add those entries into the linker white list for mscorlib which
doesn't change immensely the resulting size of the linked mscorlib (~4kB extra).