[Xamarin.Android.Build.Tasks] Unable to run Multi-Dex
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=51356

Commit a967b24c added a PROGUARD_HOME environment variable when
running Proguard and MultiDex. However it does break things on
windows. The problem is down to a bug in the Google proguard.bat
file where a " is in the wrong place. This results in the
following error.

	Unable to access jarfile C:\Program Files (x86)\Android\android-sdk\tools\proguard"\lib\proguard.jar -injars 'C:\Program'

Note the " in the wrong place. This does NOT occur if we allow
the android tooling to find proguard itself rather than specifying
a path (via the environment variable)

This commit removes the variable from the Windows
environment this seems to fix the issue.