[Xamarin.Android.Build.Tasks] Fix the use of Aot on Windows.
When running on windows if your ndk is in a path with spaces you
are in trouble. This is because the use of quoted paths causes the
cross compiler problems on windows. Specifically when dealing with
the LD_FLAGS that we pass to the compiler. It cannot handle the following

	ld_flags="foo";"bar"

because for some reason the quotes get messed up or stripped off.
We only quote paths which have spaces in them so if your ndk is in
a path this does not have spaces you will be fine.

One solution was to quote the entire --aot flag

	"--aot=<stuff>"

But that causes issues because the parse in the cross compiler was not
handling spaces correctly in that either. We ended up with errors like

	"Cannot find Path c:\Program"

So that was not an option.

BTW Visual Studio by default installs the ndk to Program Files (x86)...

oops.

This commit reworks the code to that on windows we user the
ShortPathName for the flags we pass to the aot compiler on windows.
This allow us to dispense with the quotes which cause us problems.
We already have a KernelEx class which contains the correct methods
so we just make use of them when we are not on a Unix based plaform.