[jni] Store MonoJavaGCBridgeInfo per-domain when running on desktop.
By enabling real shadow copying of assemblies for each domain created, we
are technically loading isolated copies of every Mono.Android.dll type in
those domains.

This ends up confusing the bridge code because the previous assumption that
bridged base type would be common among domains is not true anymore. This
ends up creating native crashes because the code is then trying to access
those class info when their associated domain has already been disposed
such as:

  SIGSEGV (0xb) at pc=0x00000001362c6e13, pid=72483, tid=60171

  Stack: [0x000070000268f000,0x000070000278f000],  sp=0x000070000278af60,  free space=1007k
  Native frames: (J=compiled Java code, j=interpreted, Vv=VM code, C=native code)
  C  [libmonosgen-2.0.dylib+0x133e13]  mono_class_setup_fields+0x63
  C  [libmonosgen-2.0.dylib+0x136136]  mono_class_init+0x386
  C  [libmonosgen-2.0.dylib+0x140a6f]  mono_class_is_subclass_of+0x1f
  C  [libmono-android.debug.so+0xa608]  gc_bridge_class_kind+0x48

Thus this commit alters the way the bridge info are stored by putting them
on the existing `MonodroidBridgeProcessingInfo` struct that is maintained
for each domain created when running on desktop. It also removes direct
references to the shared `mono_java_gc_bridge_info` variable instead
retrieving this information indirectly via `get_gc_bridge_index`.

For all of this to work properly, it relies on the fact that the GC bridge
routines are called from the correct domain for the object queried which
is retrieved via `mono_domain_get`.