[Mono.Android] Add environment variable to selectively disable Android's SynchronizationContext.
Under designer context, a main looper is created and destroyed for each render. This means that
an asynchronous operation that tries to marshal back to the main thread after it has finished
executing will never behave properly as it may happen while in a limbo state where no current
looper is present.

The result is that Android views that are coded like this:

	async void Initialize ()
	{
		Text = await GetLabelAsync ();
	}

	async Task<string> GetLabelAsync ()
	{
		await Task.Delay (5);
		return "Foo";
	}

Will fail because upon calling `Android.App.SyncContext.Post` the main looper will be gone
causing an NRE at runtime. This has the consequence of killing the designer rendering process
seemingly randomly after a layout is rendered.

There is no proper way to fix this on our side, it's just a side-effect of how the designer
work and it's up to the developer to use something like the design-time property `IsInEditMode`
to not utilize asynchronous methods when their views are being rendered in a designer context.

However because Xamarin.Forms specifically (for the XF previewer) is adept at doing those kind
of things, this commit add a workaround which allows us to disable the Android synchronization
context so that every `await` behave as if `.ConfigureAwait (false)` had been called on the
awaitable expression. Note that doing this implies the continuation will be executed on the same
run context than what precedes it which may create other issues but it's better than actively
ignoring those continuations and risking e.g. a deadlock in user code.