[Xamrain.Anroid.Build.Tasks] Fix a problem with creating zip files with a large number of files.
Currently we are hitting an issue where the following error is
produced.

	System.IO.IOException: Too many open files

This is because of the way in which libzip works. The creation of
the zip happens when the archive is closed. Until that point
file handles are kept open.. So if you keep adding files at
some point you are going to run out of file handled!

So this commit adds a ZipArchiveEx class which wraps the
creation of the archives. Every 50 files the archive is closed
and then reopend so more files can be added.
This works around the current limitation in the libzip library
itself. It would be really handy if libzip had a Flush method,
but unfortunately that is not available at this time.