[jni] Specify a directory to match from when shadow-copying assemblies.
Supersede https://github.com/xamarin/xamarin-android/pull/398

This commit adds an argument to the `monodroid_create_appdomain`
function to set which directory should have the assemblies it
contains shadow-copied by the runtime.

Since in debug mode XA is setup with two override directories
that contains the assemblies to load, we can use this to our
advantage to assign one of those directory as containing assemblies
that should be shadow-copied and another where they won't be.

From the caller side, assemblies that are considered unique
for the installed XA (essentially anything in the 1.0 moniker)
can be ignored for shadow-copy.

In addition, due to the current limitations expressed in the above
PR link with the GC bridge and multiple loaded Mono.Android.dll
assemblies, we can also special-case for now that Mono.Android.dll
should be loaded from one location and thus shared across all
appdomains.