[msbuild] fix StubApplication access to private fields.
Android N seems to have changed ResourceManager internals so that now
there is ResourcesImpl class that is to handle several resource accesses.
That unfortunately impacts our fast deployment implementation, which
works like InstantRun and faces the same kind of problems; it needs to
access nonpublic framework internals.

To catch up and workaround the issue, now access to those fields are
two folds for AssetManager and optional for resources.

Resource access on Android N may need further fixes, but so far it
does not crash at startup anymore.