[Xamarin.Android.Build.Utilities] Allow the use of MONO_ANDROID_PATH on windows.
We need to be able to support custom paths on wwindows as
well as Mac/Linux. However the directory structure is slightly
different. Windows does not have the tools in lib/mandroid but
in the root of the path.

This change allows us to use MONO_ANDROID_PATH on windows by
validating the Root of the path provided as well.

There are also changes to the location of the Reference Assemblies
which needed to be taken into account.

There was also a bug where if the AndroidSdk was created BEFORE
the MonoDroidSdk it was creating a MonoDroidsdk automatically.
This ment that we would be unable to update the paths.

This commit fixes those issues too.