[msbuild] add AndroidBuildApplicationPackage property.
It works as an alternative to /t:SignAndroidPackage.

I was inspired by Clancey for this change:

> On iOS we pass /p:BuildIpa=true into the xbuilds parameters and
> it works for iOS.  IS it possible to do something similar so
> Android can also build from an SLN?

That's quite convenient. In other words, it had been quite inconvenient
that we cannot get *.apk without having to find Android project file.

That's also bringing awkward inconsistency in Mobile Center.

Having an additional property fixes it, and it is feature parity with iOS.