[Xamarin.Android.Build.Tasks] Allow overriding $(AutoUnifyAssemblyReferences) (#413)
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=51831

According to the bug report, AutoUnity on `ResolveAssemblyReference` is
causing the build to lock up. At this time the user is unable to alter
the behaviour to work around the issue. This commit updates the targets
to allow users to override `$(AutoUnifyAssemblyReferences)` by passing it
as a parameter to the build.

	msbuild foo.csproj /t:Build /p:AutoUnifyAssemblyReferences=False