[monodroid] Don't use getpid() when opening netlink session (#417)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=41860

We have previously used `getpid()` when opening a netlink session  but it turns
out that WebView/Chromium does the same and keeps the session open. However,
there can only be one session with the same PID and so opening another one after
WebView instance is created in the application makes the netlink socket bind
fail with EADDRINUSE. Setting it to 0 will cause the kernel to assign some PID
that's unique and valid instead.