[Xamarin.Android.Build.Tasks] Remove use of $(MSBuildRuntimeType) (#420)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=51356

It turns out that `$(MSBuildRuntimeType)` is empty by default on windows.
This means that all of our conditional code which is only supposed
to run on windows...is not running at all.

Rather than using `$(MSBuildRuntimeType)` to figure out which platform
we are running on we just use the `Environment.OSVerson.Platform`
instead. If we are running Windows we current do not export the
`PROGUARD_HOME` environment variable. On Linux or Mac we should.