[build] Fix oss-xamarin.android*.zip generation
Commit 10daaea9 was a valiant effort, Worked On My Machine™, and
[broke the build][0]:

	zip -r "oss-xamarin.android_v7.1.99.87_Darwin-x86_64_HEAD_10daaea.zip" \
			 \

	zip error: Nothing to do! (oss-xamarin.android_v7.1.99.87_Darwin-x86_64_HEAD_10daaea.zip)
	make: *** [oss-xamarin.android_v7.1.99.87_Darwin-x86_64_HEAD_10daaea.zip] Error 12
	Build step 'Execute shell' marked build as failure

The problem is `$(wildcard)` evaluation time: it's evaluated when
`BuildEverything.mk` is *read*, **not** when the target is executed.

Consequently, e.g. `$(wildcard $(ZIP_OUTPUT_BASENAME)/bin/Debug)`
expands to the empty string, because -- at the time that
`BuildEverything.mk` is *read* -- `$(ZIP_OUTPUT_BASENAME)` does not
yet exist!

The result is that we're telling `zip` to do *absolutely nothing*,
which it apparenlty doesn't like, and errors accordingly.

The fix? Don't use `$(wildcard)`. Of course, we were using
`$(wildcard)` because some of the referenced files might not exist --
we want to start running `make package-oss` on PR builds, which will
only build the Debug configuration, so we shouldn't error out when
`bin/Release` doesn't exist -- so we need a way to get the files which
*do* exist, *at the time the target executes*.

The solution? **ls**(1), of course:

	zip -r "$(ZIP_OUTPUT)" \
		`ls -1d $(_BUNDLE_ZIPS_INCLUDE) 2>/dev/null` \
		--exclude `ls -1d $(_BUNDLE_ZIPS_EXCLUDE) 2>/dev/null`

The downside here is we can't have `--exclude` values for *each*
expanded `$(_BUNDLE_ZIPS_EXCLUDE)` value. Fortunately, that isn't
required -- the **zip**(1) man page gives examples where globs listed
after the `--exclude`:

> Multiple patterns can be specified, as in:
>
>     zip -r foo foo -x \*.o \*.c

Having a single `--exclude` should thus be fine.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/239/console