[build] Improve branch name construction
Two important parts of the build system rely on the git branch name
for "external" consumption:

* `make package-oss` uses the branch name in the `oss-xamarin.android`
    filename; see commit 1eebbe33.

* `src/Mono.Android` emits the branch name into the generated
    `Properties\AssemblyInfo.cs` within the
    `[assembly:AssemblyInformationalVersion]` value;
    see commit b620689d

Both of these uses used `git rev-parse --abbrev-ref HEAD` which
works...so long as you're *not* on a detached branch.

The problem is that when building on Jenkins, the build *is* on a
detached branch, as part of the build shenanigans involves
`git checkout -f COMMIT`. When on a detached branch,
`git rev-parse --abbrev-ref HEAD` *always* returns `HEAD`, meaning we
always think that the branch name is `HEAD`.

This results in filenames like:

	oss-xamarin.android_v7.1.99.86_Darwin-x86_64_HEAD_2ca4a32.zip

`HEAD` is not entirely useful there.

Instead of using `git rev-parse`, we'll instead use
`git branch --contains HEAD`. This (more?) reliably provides a useful
branch name, even on detached branches.