[Xamarin.Android.Build.Utilities] Get registry values from a custom registry key support for VS2017 SXS
VS2017 introduce the possibility to install more than one instance.
We need to maintain the preferred path (javasdk, androidsdk, androidndk)
of XA isolated by VS installation instance.

If user has VS2015 and VS2017 (maybe more than 1 installed) the path
must be the reference to the installation path of each product
(javasdk, androidsdk, androidndk) for each version.

XA store it in the registry for whole Xamarin.
We needs to define an environment variable to set the regkey to be
used at the moment of resolve/store the preferred path.

VS and XA (msbuild targets), this environment variable is assigned
with a string that represents the registry key based on VS installation
id for VS2017 (XAMARIN_ANDROID_REGKEY=SOFTWARE\Xamarin\VisualStudio_{id}\Android).

VS2015 and previous continues working with the general regkey defined in XA (SOFTWARE\Novell\Mono for Android).