[build] Revert improved branch name construction
Commits badec2b4 and 3f41f7e7 *attempted* to improve determination of
branch names when running on Jenkins, but they failed;
`git branch --contains HEAD` also failed to provide any meaningful
branch names when building on Jenkins. The end result of which was
that, instead of meaingles `HEAD` being detected as the branch name,
`""` (the empty string) was being detected instead!

This is not better. :-(

Improved reading of the Jenkins documentation (thanks @alanmcgovern!)
directed us to the [Check out to specific local branch][0] option,
which looks like it'll allow `git rev-parse` to work as desired.

Revert commits badec2b4 and 3f41f7e7 so that `git rev-parse` is once
again used to determine branch names.

[0]: http://stackoverflow.com/questions/39297783/detached-head-w-jenkins-git-plugin-and-branch-specifier