[Xamarin.Android.Build.Tasks] Revert the creation of the sdks.cache
Context https://bugzilla.xamarin.com/show_bug.cgi?id=52451
	https://bugzilla.xamarin.com/show_bug.cgi?id=32578

The sdks.cache was introduced a while ago to speed up builds.
However with VS2017 this only seems to cause problems.

With changes to androidtools the ResolveSdks task only takes
around 300 ms to run. So to save us some head aches in the future
we should remove the caching of the sdks.