[Xamarin.Android.Build.Tasks] Properties Panel in VS not updating (#440)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=51604

Custom control "CardView" properties are not getting appearing
in Properties Panel in VS. (Take 2!)

Adding a Nuget package to the project in the IDE does not
seem to be causing a build. It appears that the build system
thinks the project file is up to date.

This commit adds a check to see if the packages.config has changed
(if it exists). If it has changed then we need to rebuild the
referencepaths.cache again to ensure we include the new assemblies
resources.