[Mono.Android] Ensure no Java network code runs on main thread (#409)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=51774

`AndroidHttpHandler` invokes the Java HTTP client to do its actual work and the
Java code must not run on the main thread lest it throws an exception. In most
cases the handler does the right job, the exception being a handful of
HttpClientConnection.Disconnect calls invoked directly in the handler code as
well as from a cancellation token callback.

This commit makes sure we don't run any of the Java code on the main thread by
wrapping the calls in a task.