[build] Auto-Provision Dependencies
The #runtime team is requesting that Xamarin.Android support
"auto-provisioning": the ability to track dependency *versions*, and
if there is a "version mismatch", *install* a dependency that
satisfies the minimum.

The primary use case is Mono itself: there is a desire to build
xamarin-android *itself* against mono/master, to see what breaks.
(Does the C# compiler break things? Does the JIT hang? etc.)

Extend the existing `@(RequiredProgram)` infrastructure to attempt to
support this:

* `%(RequiredProgram.MinimumVersion)`: The minimum program version
    that is supported.
* `%(RequiredProgram.MaximumVersion)`: The maximum program version
    that is supported.
* `%(RequiredProgram.CurrentVersionCommand)`: A command to execute to
    obtain the current program version.
    Defaults to `%(RequiredProgram.Identity) --version`.
* `%(RequiredProgram.MinimumUrl)`: URL to download "something" to use
    to obtain a version of the program that satisfies
    `%(MinimumVersion)`.
* `%(RequiredProgram.Install)`: Command to execute which installs the
    minimum version of the program.

`%(RequiredProgram.CurrentVersionCommand)`,
`%(RequiredProgram.MinimumUrl)`, and `%(RequiredProgram.Install)` can
all be "overridden" for platform-specific values by prefixing the
value with `$(HostOSName)` and/or `$(HostOS)`. Thus, macOS-specific
values can be stored into e.g. `%(RequiredProgram.DarwinMinimumUrl)`
and `%(RequiredProgram.DarwinInstall)`.

Auto-Provisioning is controlled by the new `$(AutoProvision)`
property, which can be provided in `Configuration.Override.props`.
It defaults to `False`.

If the `$(AutoProvisionUsesSudo)` MSBuild property is True, then
`sudo` is used to invoke all install commands.