[Mono.Android] Roslyn String Interpolation Fixes
Roslyn doesn't like a "bar" `:` within string interpolations, as
[the `:` is used to separate the value from the string format][0]:

	// Note `:hh`, providing the format string "hh"
	Console.WriteLine($"Name = {name}, hours = {hours:hh}")

This causes problems with using the ternary operator:

	Console.WriteLine($"Name={name != null ? name : "unset"}");
	// error CS8076: Missing close delimiter '}' for interpolated expression started with '{'.

As it happens, this is (more or less) what
`AndroidClientHandler.HandlePreAuthentication()` was doing.
`mcs` allowed this; Roslyn does not.

...which is not to say that you can't use the ternary operator within
string interpolations; you just need to wrap in parens:

	Console.WriteLine($"Name={(name != null ? name : "unset")}");

Fix `AndroidClientHandler.cs` compilation under Roslyn.

Thanks to @borgdylan for pointing this out.

[0]: https://msdn.microsoft.com/en-us/library/dn961160.aspx