[build] `make prepare-deps` runs first
The PR builder is wedged. It needs a specific mono, but it can't
install mono because `$(call GetPath,...)` returns a bad value [^1].

Run `make prepare-deps` *first*, so that we can install a proper mono
before we try anything else.

Additionally, use "chained dependencies" so that the `prepare-*`
targets are `make -j`-friendly.

Finally, because the `$(AutoProvision)` logic installs Mono 4.8 as a
baseline (by default), the [xamarin-android-msbuild][1] job is now
building with Mono 4.8, whereas previously it used Mono 4.6.
[This *appears* to have broken things][2], because the
`_GenerateFrameworkList` target in `src/Mono.Android` isn't being
executed when it should be, resulting in:

	error MSB3644: The reference assemblies for framework "MonoAndroid,Version=v7.1" were not found.

Change the default "minimal" Mono macOS .pkg to Mono 4.6, as that's
"known good". Mono 4.8 support will be in a separate commit.