[Xamarin.Android.Build.Tasks] NDK r12b+ is not supported
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=48678

We had hardcoded the min api levels for each arch. However
google removed api 4 from the ndk platforms which causes this
error. We also did not raise the error we got. We kinda just
logged it..

This commit adds support for searching for the supported platform
instead of just blindly using one which may not exist.

We also changed the system to raise an Error if we get an exception
while running the build.