[Xamarin.Android.Build.Tasks]  AndroidExplicitCrunch does not parallelize calls to aapt (#462)
Fixed https://bugzilla.xamarin.com/show_bug.cgi?id=36079

When building with alot of resources and using $(AndroidExplicitCrunch)
we need to copy/process/copy all the resource files. Many of these
files live in different directories and therefore can be processed in
parallel.

This commit reworks the Crunch task to make use of the AsyncTask so that
each group of files can be processed at the same time.