[mono-runtimes] Install $prefix/lib/mandroid/cil-strip.exe
`cil-strip.exe` is used by the `<CilStrip/>` task in
`Xamarin.Android.Build.Tasks.dll`, and is just a renamed copy of
`mono-cil-strip.exe`.

Additionally, update the `_InstallMonoDoc` target to always `<Touch/>`
the created files. This is to keep the `_InstallMonoDoc` task from
constantly re-executing for me when running `xbuild /t:ForceBuild`:

	Target _InstallMonoDoc needs to be built as input file
	'.../xamarin-android/external/mono/mcs/class/lib/net_4_x/mdoc.exe'
	is newer than output file
	'../../bin/Debug//lib/mandroid/monodoc.dll.config'