[Xamarin.Android.Build.Tasks] Fixed DesignTime build so it does not always run. (#468)
There was a bug where the build.props file was not created as
part of a design time build. This usually occurs when the IDE's
call "UpdateAndroidResources". The reason is many of the Targets
have the build.props as part of their Inputs. So if an Input
does not exist then the entire target will run.

The fix is to make sure that the _CreatePropertiesCache target is
called as part of the designer builds. This will make sure that
build.props is created even if a design time build is run first.