[Xamarin.Android.Build.Tasks] AAPT packaging error ignored (#471)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=53054

The regex used to detect android tool errors was incorrect.
It looks like newer versions of the android tooling use brakets
rather than colons to delimit the file and line numbers.

This commit reworks the regex so it works with both the old
format and the new format. It also reworks the code so we do
not have a duplicate regex in Aapt. The existing regex in
AndroidToolTask has been made internal so that other classes
can use it. Normally we would derive from AndroidToolTask, but
in the case of Aapt we wanted to derive from AsyncTask so we could
run multiple calls in parallel.

This commit also fixes a bug where if Aapt failed for what ever reason
the parellel tasks were not cancelled.