[Xamarin.Android.Build.Tests] Make sure the build.logs are included in the output.
One of the problems with the jenkins build is the logs from the
failed tests are not available on the website. This commit makes
sure that any tests that did not pass have the build.log information
written to the TestContext.Out. On jenkins this is setup to
write to a file in the bin\Test$(Configuration) directory.
The Makefile has also been updated to write this information
to the output. This will allow the data to appear in the raw
output log.