[Xamarin.Android.Build.Tasks] detect OS changes.
If a user shares a project between Mac and Windows they
will probably get some errors. This is mainly because
the various cache's that are in place will contain
invalid paths.

This commit adds the $(OS) to the build.props file which
will allow us to detect if the OS changed.