[build] `make leeroy` should build Xamarin.Android.sln
When building some internal repositories, we were encountering an
error due to an "inconsistent build environment":

	System.IO.FileNotFoundException: Could not load file or assembly
	'Xamarin.Android.Cecil, Version=0.10.0.0, Culture=neutral, PublicKeyToken=0738eb9f132ed756'
	or one of its dependencies

Lots of head-scratching later, we still don't know what was *removing*
`external/Java.Interop/bin/Release/Xamarin.Android.Cecil.dll`, but to
some extent, *it doesn't matter* what removed it.

What matters is the sanity and consistency of the build tree.

The problem with `make leeroy` is that it tried to be "too smart for
its own good", just depending on *select* build artifacts, to prompt
`make` invocations to build those artifacts, *with the implication*
that *all the other* artifacts we needed would *also* be built.

This is sufficient for `make jenkins`, and similar environments where
you always start from a clean slate and build everything.

This is *not* sufficient for a developer tree. Because only "select"
outputs are known to make, *all other* outputs are ignored -- and
there are lots! -- and these other artifacts can be *important*.

We need a "known good" build tree. The only way to get that is to
build the top-level Xamarin.Android.sln, and `make leeroy` wasn't
ensuring that.

Update `make leeroy` to ensure that `Xamarin.Android.sln` is *always*
built. Worst case, it'll length build times.