Grab response content on redirects with AllowAutoRedirect == false
Currently when automatic redirection is disabled and a request is made that
receives a redirect code in response we ignore the content and the headers. This
is incorrect as such a response is a valid one and the client might want to
handle redirection on their own, using the information contained in response
headers.

This commit makes sure that all the headers, as well as cookies, are present in
the response message returned to the caller.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=50587